Fix #1164: remove offending headers (#1169)
Exploring a possible fix that fixes
https://github.com/pulumi/pulumi-cloudflare/issues/1164

We have MapNestedAttribute "headers":


https://github.com/cloudflare/terraform-provider-cloudflare/blob/72a74193561eaaf004a9aa21499c44ca729e8bd3/internal/services/ruleset/schema.go#L348

Before:

```
.rules[0].actionParameters[0].headers[0]
```

After:

```
.rules[0].actionParameters.headers{"X"}
```

Unfortunately even for users that do not set headers the empty
header-list fails to parse under the map interpretation. This change
removes empty headers lists so that at least these programs succeed.

It is difficult to argue that this should be a bridge fix since it seems
the schema change calls for an explicit migration.