Fix ruleset rules migration (#1244)
The `rules` property within the `actionParameter` property was changed
from `map[string]string` into `map[string][]string` upstream but without
a migration. This adds that missing migration so users don't hit issues
with it.

I have not managed to actually create a resource with the properties in
question but have reproduced the issue with a state a user provided.
I've added the test here.

fixes https://github.com/pulumi/pulumi-cloudflare/issues/1213