Fix ruleset rules for older providers (#1285)
The `Ruleset` resource had some type changes and no corresponding
migration. We fixed this in
https://github.com/pulumi/pulumi-cloudflare/pull/1244 but it only
covered somewhat recent provider versions due to the resource version
guard (post 2022 or so, v 4.10.0).

This PR removes the version guard and does the migration only based on
the types of the properties in the state. This should cover all provider
versions.

I've also added tests for a state from v4 and v6.

fixes https://github.com/pulumi/pulumi-cloudflare/pull/1213