Remove an unnecessary test module (#1418)
The presence of this module blocks automated bridge rollouts because it
requires us to manually tidy on each update. But it's not necessary to
treat this as a standalone module -- it's just a test in our repo that
we want to run. It can and should belong to our main module.

(As a reminder, modules essentially tell Go to treat subdirectories as
if they are separate repositories. That almost always introduces
unnecessary coordination and dependency complexity.)

Related https://github.com/pulumi/pulumi/issues/21086 -- as a workaround
we can skip the install step.